Clean up some inheritance for existing requests
